requirements.txt
pandas
sqlalchemy
mysql-connector-python
matplotlib
seaborn
scikit-learn
streamlit
python-dotenv